refactor: 合并 hexToRgba + 删除三个未消费的调度服务函数 (Bugbot Low x2)
Bugbot Low #1 (重复)：WeeklyPosterModal 里新加的 hexToRgba 重复了 lib/
  themeComputed.ts 已有的同名函数。差异是新版多了 #RGB 简写支持 + 非法兜底，
  老版（fragile）会在 hex.length < 7 时 NaN 输出 "rgba(NaN,NaN,NaN,...)"。
  合并：把 robust 版搬进 themeComputed.ts 替换原 fragile 版，删 modal 里的
  本地副本，import { hexToRgba } from '@/lib/themeComputed'。原 themeSystem
  56 个单测全部通过（兼容）。

Bugbot Low #2 (死代码)：listWorkflowSchedules / updateWorkflowSchedule /
  deleteWorkflowSchedule 这三个 service 函数 + 三个 Contract type + 三个
  Real 实现没有任何调用方。当前只有 createWorkflowSchedule 在 AutoPublishDialog
  里用，剩余 CRUD 是为将来"调度管理 UI"提前预留 —— 但 CLAUDE.md 规则禁
  止"为假想未来需求设计"。删干净，要用时再加。
