+++ prd-admin/src/services/index.ts
@@ -1544,10 +1544,7 @@ import type {
   TestRunCapsuleContract,
   GetChatHistoryContract,
   AiFillParametersContract,
-  ListSchedulesContract,
   CreateScheduleContract,
-  UpdateScheduleContract,
-  DeleteScheduleContract,
 } from '@/services/contracts/workflowAgent';
 export type { Workflow, WorkflowVariable, WorkflowExecution, WorkflowSchedule } from '@/services/contracts/workflowAgent';
 import {
@@ -1571,10 +1568,7 @@ import {
   testRunCapsuleReal,
   getChatHistoryReal,
   aiFillParametersReal,
-  listWorkflowSchedulesReal,
   createWorkflowScheduleReal,
-  updateWorkflowScheduleReal,
-  deleteWorkflowScheduleReal,
 } from '@/services/real/workflowAgent';
 
 export const listWorkflows: ListWorkflowsContract = withAuth(listWorkflowsReal);
@@ -1600,10 +1594,7 @@ export { chatWorkflowReal as chatWorkflow } from '@/services/real/workflowAgent'
 export { analyzeExecutionReal as analyzeExecution } from '@/services/real/workflowAgent';
 export { validateTapdCookie } from '@/services/real/workflowAgent';
 export const aiFillParameters: AiFillParametersContract = withAuth(aiFillParametersReal);
-export const listWorkflowSchedules: ListSchedulesContract = withAuth(listWorkflowSchedulesReal);
 export const createWorkflowSchedule: CreateScheduleContract = withAuth(createWorkflowScheduleReal);
-export const updateWorkflowSchedule: UpdateScheduleContract = withAuth(updateWorkflowScheduleReal);
-export const deleteWorkflowSchedule: DeleteScheduleContract = withAuth(deleteWorkflowScheduleReal);
 
 // 数据迁移服务
 import type {
